requirements.txt
# Data fetching
yfinance==0.2.51
nsepython==2.9
pandas==2.2.3
numpy==1.26.4
requests==2.32.3

# Database
# sqlite3 is built into Python — no install needed

# Scheduling / environment
python-dotenv==1.0.1
pytz==2024.1

# Telegram bot (Phase 3)
python-telegram-bot==20.8

# Dashboard (Phase 4)
streamlit==1.35.0
plotly==5.22.0

# Utilities
tabulate==0.9.0
loguru==0.7.2
